Carpeta media

+++ src/TADs/Queue/MyPriorityQueue.java
@@ -1,5 +1,5 @@
 package TADs.Queue;
 
 public interface MyPriorityQueue<T> extends MyQueue<T>{
-    public void enqueueWithPriority(T element, int prioridad);
+    public void enqueueWithPriority(T element, double prioridad);
 }
